+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes that happened in releases
 * Support exposing additional (computed) GraphQL fields in Entity Views and support ignoring types and fields
 * Support marking GraphQL fields and field types as non-null in Entity Views
 * Add `ManagedViewType#renderSecondaryMappings` implementation that accepts `FromBuilder`
+* Introduce `PredicateBuilder`, `PredicateAndBuilder` and `PredicateOrBuilder` as abstraction for `WHERE`, `HAVING` and `ON` clauses
 
 ### Bug fixes